third_party/WebKit/LayoutTests/fast/html/imports/import-insert.html

<!DOCTYPE html>

<script>
    jsTestIsAsync = true;
</script>

<script src="../../../resources/js-test.js"></script>
<script src="resources/import-helpers.js"></script>

<script>
debug("It shouldn't be loaded before insertion.");

var placeholder = createPlaceholder();
var toInsert = document.createElement("link");
toInsert.setAttribute("rel", "import");
toInsert.setAttribute("href", "resources/hello.html");

placeholder.onload = function() {
    shouldBeNull("toInsert.import");
    debug("It should be loaded once it is inserted.");
    document.head.appendChild(toInsert);
    toInsert.onload = function() {
        shouldBe("toInsert.import.querySelector('h1').innerHTML", "'Hello'");
        finishJSTest();
    };
};
</script>
