third_party/WebKit/LayoutTests/fast/html/imports/import-insert-expected.txt

It shouldn't be loaded before insertion.
PASS toInsert.import is null
It should be loaded once it is inserted.
PASS toInsert.import.querySelector('h1').innerHTML is 'Hello'
PASS successfullyParsed is true

TEST COMPLETE
